partially removed opengraph support for youtube videos.
if a video provides the url for a flash stream it will be used by default by websites like facebook, making it hard for people access blog
posts from facebook, instead of watch the videos directly on facebook. It just fits well for video sites, like youtube itself, where the video
is the most important content of the page.

added just the thumbnail url to the images array.